pyproject.toml
[tool.poetry]
name = "matematiikkavalmennusbot"
version = "0.1.0"
description = ""
authors = ["Jouni K. Seppänen <jks@iki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["all"], version = "^20.4"}
feedparser = "^6.0.10"
tomlkit = "^0.12.1"
filelock = "^3.12.2"


[tool.poetry.group.dev.dependencies]
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
